feat: merge Oracle MCP functionality into upstream structure
- Merge Oracle MCP tool resolution from routes/virtual_assistants.py into app/api/v1/virtual_agents.py
- Preserve all Oracle MCP features: tool resolution, sampling strategies, model optimizations
- Update API endpoints to use LlamaStack integration with dual storage strategy
- Fix backend structure to match upstream app/api/v1/ organization
- Remove old routes/ directory structure
- Maintain compatibility with upstream main branch while preserving Oracle MCP capabilities

# Conflicts:
#	backend/app/api/v1/virtual_agents.py
